Working parser v.2
- separated code into multiple files for readability
